+++ src/test/java/chess/domain/chessBoard/SpaceTest.java
@@ -1,6 +1,5 @@
-package chess.chessBoard;
+package chess.domain.chessBoard;
 
-import chess.domain.chessBoard.Space;
 import chess.domain.piece.Color;
 import chess.domain.piece.EmptyPiece;
 import chess.domain.piece.Pawn;
