+++ src/test/java/chess/domain/piece/KingTest.java
@@ -1,10 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
 import chess.domain.chessBoard.Space;
-import chess.domain.piece.Color;
-import chess.domain.piece.EmptyPiece;
-import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
 import chess.domain.position.File;
 import chess.domain.position.Position;
 import chess.domain.position.Rank;
@@ -16,37 +12,37 @@ import java.util.List;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-public class QueenTest {
+public class KingTest {
 
     @Test
-    @DisplayName("상하좌우로 칸 수 제한 없이 움직일 수 있다.")
-    void should_move_up_down_left_right_unlimited() {
-        Piece piece = new Queen(Color.WHITE);
+    @DisplayName("상하좌우로 한 칸 움직일 수 있다.")
+    void should_move_up_down_left_right() {
+        Piece piece = new King(Color.WHITE);
         Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.a, Rank.THREE));
+        Space space2 = new Space(new EmptyPiece(), new Position(File.a, Rank.TWO));
 
         space1.movePiece(space2, List.of(space1, space2));
 
-        assertThat(space2.pieceCharacter()).isEqualTo("q");
+        assertThat(space2.pieceCharacter()).isEqualTo("k");
     }
 
     @Test
-    @DisplayName("대각선 방향으로 칸 수 제한 없이 움직일 수 있다")
-    void should_move_diagonal_unlimited() {
-        Piece piece = new Queen(Color.WHITE);
+    @DisplayName("대각선 방향으로 한 칸 움직일 수 있다")
+    void should_move_diagonal() {
+        Piece piece = new King(Color.WHITE);
 
         Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(new EmptyPiece(), new Position(File.c, Rank.THREE));
+        Space space2 = new Space(new EmptyPiece(), new Position(File.b, Rank.TWO));
 
         space1.movePiece(space2, List.of(space1, space2));
 
-        assertThat(space2.pieceCharacter()).isEqualTo("q");
+        assertThat(space2.pieceCharacter()).isEqualTo("k");
     }
 
     @Test
-    @DisplayName("상하좌우와 대각선외에는 움직일 수 없다")
-    void should_not_move_not_upDownLeftRight_not_diagonal() {
-        Piece piece = new Queen(Color.WHITE);
+    @DisplayName("두 칸 이상 움직일 수 없다")
+    void should_not_move_more_than_two_space() {
+        Piece piece = new King(Color.WHITE);
 
         Space space1 = new Space(piece, new Position(File.a, Rank.ONE));
         Space space2 = new Space(new EmptyPiece(), new Position(File.b, Rank.THREE));
@@ -59,13 +55,13 @@ public class QueenTest {
     @Test
     @DisplayName("이동할 위치에 상대 말이 있고 잡을 수 있으면 이동할 수 있다")
     void should_move_when_target_space_has_other_color_piece_and_catchable() {
-        Piece piece1 = new Queen(Color.WHITE);
-        Piece piece2 = new Queen(Color.BLACK);
+        Piece piece1 = new King(Color.WHITE);
+        Piece piece2 = new King(Color.BLACK);
         Space space1 = new Space(piece1, new Position(File.a, Rank.ONE));
-        Space space2 = new Space(piece2, new Position(File.c, Rank.THREE));
+        Space space2 = new Space(piece2, new Position(File.b, Rank.TWO));
 
         space1.movePiece(space2, List.of(space1, space2));
 
-        assertThat(space2.pieceCharacter()).isEqualTo("q");
+        assertThat(space2.pieceCharacter()).isEqualTo("k");
     }
 }
