+++ src/test/java/chess/domain/piece/PieceTest.java
@@ -1,7 +1,5 @@
-package chess.piece;
+package chess.domain.piece;
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Piece;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
