+++ src/test/java/chess/domain/piece/TestPiece.java
@@ -1,8 +1,6 @@
-package chess.piece;
+package chess.domain.piece;
 
 
-import chess.domain.piece.Color;
-import chess.domain.piece.Piece;
 import chess.domain.position.Position;
 
 public class TestPiece extends Piece {
