+++ src/test/java/chess/domain/position/FileTest.java
@@ -1,7 +1,5 @@
-package chess.position;
+package chess.domain.position;
 
-import chess.domain.position.File;
-import chess.domain.position.FileDifference;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
