+++ src/test/java/chess/domain/position/PositionTest.java
@@ -1,8 +1,5 @@
-package chess.position;
+package chess.domain.position;
 
-import chess.domain.position.File;
-import chess.domain.position.Position;
-import chess.domain.position.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
